+++ src/vite-env.d.ts
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: MIT OR LicenseRef-Moria-Angband
 /// <reference types="vite/client" />
+/// <reference types="vite-plugin-pwa/client" />
 
 interface ImportMetaEnv {
   readonly VITE_GIT_REVISION: string;